+++ app/client/packages/rts/src/ee/server.ts
@@ -1,0 +1,1 @@
+import "../ce/server.ts";
