+++ app/server/appsmith-server/src/main/java/com/appsmith/server/configurations/CommonConfig.java
@@ -64,15 +64,6 @@ public class CommonConfig {
     @Value("${disable.telemetry:true}")
     private boolean isTelemetryDisabled;
 
-    @Value("${appsmith.observability.tracing.detail.enabled:false}")
-    private boolean tracingDetail;
-
-    @Value("${appsmith.observability.metrics.detail.enabled:false}")
-    private boolean metricsDetail;
-
-    @Value("${appsmith.observability.metrics.interval.millis:60000}")
-    private int metricsIntervalMillis;
-
     private List<String> allowedDomains;
 
     private String mongoDBVersion;
