+++ app/server/appsmith-server/src/main/resources/application-ce.properties
@@ -1,0 +1,111 @@
+spring.application.name=appsmith-server
+server.port=${PORT:8080}
+# Allow the Spring context to close all active requests before shutting down the server
+# Please ref: https://docs.spring.io/spring-boot/docs/2.3.0.RELEASE/reference/html/spring-boot-features.html#boot-features-graceful-shutdown
+server.shutdown=graceful
+server.max-http-request-header-size=16KB
+
+spring.lifecycle.timeout-per-shutdown-phase=20s
+
+spring.profiles.active=${ACTIVE_PROFILE:production}
+
+appsmith.db.url=${APPSMITH_DB_URL:${APPSMITH_MONGODB_URI}}
+# This property allows us to override beans during testing. This is useful when we want to set different configurations
+# and different parameters during test as compared to production. If this property is disabled, some tests will fail.
+spring.main.allow-bean-definition-overriding=true
+spring.data.redis.repositories.enabled=false
+
+#spring.main.web-application-type=reactive
+
+# This property allows the server to run behind a proxy server and still resolve all the urls correctly
+server.forward-headers-strategy=NATIVE
+
+spring.data.mongodb.auto-index-creation=false
+spring.data.mongodb.authentication-database=admin
+# Ensures that the size of the request object that we handle is controlled. By default it's 212KB.
+spring.codec.max-in-memory-size=150MB
+# The value is same as appsmith.codec as both these values serves same purpose.
+spring.webflux.multipart.max-in-memory-size=${APPSMITH_CODEC_SIZE:150}MB
+appsmith.codec.max-in-memory-size=${APPSMITH_CODEC_SIZE:150}
+
+# Log properties
+logging.level.root=info
+logging.level.com.appsmith=debug
+logging.level.com.external.plugins=debug
+logging.pattern.console=[%d{ISO8601, UTC}] [%t] requestId=%X{X-Appsmith-Request-Id} userEmail=%X{userEmail} traceId=%X{traceId} spanId=%X{spanId} - %m%n
+
+#Spring security
+spring.security.oauth2.client.registration.google.client-id=${APPSMITH_OAUTH2_GOOGLE_CLIENT_ID:missing_value_sentinel}
+spring.security.oauth2.client.registration.google.client-secret=${APPSMITH_OAUTH2_GOOGLE_CLIENT_SECRET:}
+spring.security.oauth2.client.provider.google.userNameAttribute=email
+spring.security.oauth2.client.registration.github.client-id=${APPSMITH_OAUTH2_GITHUB_CLIENT_ID:missing_value_sentinel}
+spring.security.oauth2.client.registration.github.client-secret=${APPSMITH_OAUTH2_GITHUB_CLIENT_SECRET:}
+spring.security.oauth2.client.provider.github.userNameAttribute=login
+
+# Accounts from specific domains are allowed to login
+# DEPRECATED, in favor of signup.allowed-domains
+oauth2.allowed-domains=${APPSMITH_OAUTH2_ALLOWED_DOMAINS:}
+
+# Segment
+segment.writeKey=${APPSMITH_SEGMENT_KEY:}
+# Is this instance hosted on Appsmith cloud?
+is.cloud-hosting = ${APPSMITH_CLOUD_HOSTING:false}
+disable.telemetry = ${APPSMITH_DISABLE_TELEMETRY:true}
+segment.ce.key = ${APPSMITH_SEGMENT_CE_KEY:}
+logging.verbose.enabled = ${APPSMITH_VERBOSE_LOGGING_ENABLED:false}
+
+# Sentry
+sentry.dsn=${APPSMITH_SERVER_SENTRY_DSN:}
+sentry.send-default-pii=true
+sentry.debug=false
+sentry.environment=${APPSMITH_SERVER_SENTRY_ENVIRONMENT:}
+
+# Redis Properties
+appsmith.redis.url=${APPSMITH_REDIS_URL}
+
+# Mail Properties
+# Email defaults to false, because, when true and the other SMTP properties are not set, Spring will try to use a
+#   default localhost:25 SMTP server and throw an error. If false, this error won't happen because there's no attempt
+#   to send an email.
+mail.enabled=${APPSMITH_MAIL_ENABLED:false}
+mail.from=${APPSMITH_MAIL_FROM:}
+mail.support=${APPSMITH_MAIL_SUPPORT:support@appsmith.com}
+reply.to=${APPSMITH_REPLY_TO:appsmith@localhost}
+spring.mail.host=${APPSMITH_MAIL_HOST:}
+spring.mail.port=${APPSMITH_MAIL_PORT:}
+spring.mail.username=${APPSMITH_MAIL_USERNAME:}
+spring.mail.password=${APPSMITH_MAIL_PASSWORD:}
+spring.mail.properties.mail.smtp.auth=${APPSMITH_MAIL_SMTP_AUTH:}
+spring.mail.properties.mail.smtp.starttls.enable=${APPSMITH_MAIL_SMTP_TLS_ENABLED:}
+admin.emails = ${APPSMITH_ADMIN_EMAILS:}
+
+# Appsmith Cloud Services
+appsmith.cloud_services.base_url = ${APPSMITH_CLOUD_SERVICES_BASE_URL:}
+appsmith.cloud_services.signature_base_url = ${APPSMITH_CLOUD_SERVICES_SIGNATURE_BASE_URL:}
+appsmith.cloud_services.template_upload_auth_header = ${APPSMITH_CLOUD_SERVICES_TEMPLATE_UPLOAD_AUTH:}
+github_repo = ${APPSMITH_GITHUB_REPO:}
+
+# Support disabling signup with an environment variable
+signup.disabled = ${APPSMITH_SIGNUP_DISABLED:false}
+signup.allowed-domains=${APPSMITH_SIGNUP_ALLOWED_DOMAINS:}
+
+# Google recaptcha config
+google.recaptcha.key.site = ${APPSMITH_RECAPTCHA_SITE_KEY:}
+google.recaptcha.key.secret= ${APPSMITH_RECAPTCHA_SECRET_KEY:}
+
+# Plugin Interface level settings
+appsmith.plugin.response.size.max=${APPSMITH_PLUGIN_MAX_RESPONSE_SIZE_MB:5}
+
+# Location env file with environment variables, that can be configured from the UI.
+appsmith.admin.envfile=${APPSMITH_ENVFILE_PATH:/appsmith-stacks/configuration/docker.env}
+
+# RTS port
+appsmith.rts.port=${APPSMITH_RTS_PORT:}
+
+appsmith.internal.password=${APPSMITH_INTERNAL_PASSWORD:}
+
+# GIT stale index.lock file valid time
+appsmith.index.lock.file.time=${APPSMITH_INDEX_LOCK_FILE_TIME:300}
+
+springdoc.api-docs.path=/v3/docs
+springdoc.swagger-ui.path=/v3/swagger
