+++ app/server/appsmith-server/src/main/resources/application-ee.properties
@@ -1,0 +1,1 @@
+spring.config.import=application-ce.properties
